Make the build config more developer-friendly.
The aim of this commit is to make the nix config usable for both the
build (nix-build) and the development (nix-shell + cabal). Now, you can
"just nix-build" as before but also "run a nix-shell" in the
subdirectories to incrementally build and test the subprojects (common,
client, server).

Important changes:

- add some cabal files & default.nix in the subdirectories

- use callCabal2nix to automatically generate nix configs from cabal
files

- rename the "shared" directory to "common" (to avoid potential
confusion with the "shared" keyword in the cabal files)

- add a nixpkgs.nix file for building a compatible nixpkgs environment

- add a section in the README.md explaining how to use the development
config